Added pimpl idiom for pcicclient
Lowering dependencies for projects including
this modul. Especially the boost dependencies.

This fixes #50

Acked-by: Christian Ege <christian.ege@ifm.com>
