+++ modules/pcicclient/include/ifm3d/pcicclient/pcicclient.h
@@ -23,8 +23,6 @@
 #include <mutex>
 #include <string>
 #include <thread>
-#include <boost/asio.hpp>
-#include <boost/system/system_error.hpp>
 #include <ifm3d/camera.h>
 
 namespace ifm3d
@@ -34,6 +32,7 @@ namespace ifm3d
    * ifm3d::Camera::Ptr, it provides unbuffered communication with
    * the PCIC interface.
    */
+   
   class PCICClient
   {
   public:
@@ -150,214 +149,9 @@ namespace ifm3d
     void CancelCallback(long callback_id);
 
   private:
-    /**
-     * Commands consist of content data surrounded by some meta data.
-     * The State enum provides information which buffer is currently
-     * used in writing to and reading from network
-     */
-    enum class State { PRE_CONTENT, CONTENT, POST_CONTENT };
-
-    /**
-     * Connects to the camera
-     */
-    void DoConnect();
-
-    /**
-     * Handles DoConnect results
-     */
-    void ConnectHandler(const boost::system::error_code& ec);
-
-    /**
-     * Reads data from network into one of the three "in" buffers
-     * depending on current reading state.
-     */
-    void DoRead(State state, std::size_t bytes_remaining = UNSET);
-
-    /**
-     * Handles DoRead results: Triggers further reads and in
-     * case an incoming message is completely received, does
-     * the callback (if existent).
-     */
-    void ReadHandler(State state, const boost::system::error_code& ec,
-                     std::size_t bytes_transferred,
-                     std::size_t bytes_remaining);
-
-    /**
-     * Returns buffer to be filled from network depending on
-     * specified reading state
-     */
-    std::string& InBufferByState(State state);
-
-    /**
-     * Writes data to network from one of the three "out" buffers
-     * depending on current writing state.
-     */
-    void DoWrite(State state,
-                 const std::string &out_content_buffer,
-                 std::size_t bytes_remaining = UNSET);
-
-    /**
-     * Handles DoWrite results: Triggers further writes and in
-     * case a request is completely sent, unblocks calling thread.
-     */
-    void WriteHandler(State state,
-                      const boost::system::error_code& ec,
-                      std::size_t bytes_transferred,
-                      const std::string& out_content_buffer,
-                      std::size_t bytes_remaining);
-
-    /**
-     * Returns buffer containing data to be written to network
-     * depending on specified writing state. (In case of state CONTENT,
-     * the specified out_content_buffer is returned.)
-     */
-    const std::string& OutBufferByState(State state,
-                                        const std::string& out_content_buffer);
-
-    /**
-     * Finds and returns the next free ticket for a command
-     */
-    int NextCommandTicket();
-
-    /**
-     * Calculates and returns next callback id
-     */
-    long NextCallbackId();
-
-  private:
-    /**
-     * Init command sequence
-     */
-    static const std::string init_command;
-
-    /**
-     * Shared pointer to the camera this PCIC client will communicate with.
-     */
-    ifm3d::Camera::Ptr cam_;
-
-    /**
-     * Cached copy of the camera IP address
-     */
-    std::string cam_ip_;
-
-    /**
-     * Cached copy of the camera PCIC TCP port
-     */
-    int cam_port_;
-
-    /**
-     * Flag indicating that client is connected.
-     */
-    std::atomic_bool connected_;
-
-    /**
-     * Flag which is used as default parameter in DoRead and DoWrite
-     * in order to indicate that a new buffer is used for read/write
-     * so that the remaining size can be initialized by buffers size.
-     */
-    static const int UNSET = -1;
-
-    /**
-     * The ASIO event loop handle
-     */
-    boost::asio::io_service io_service_;
-
-    /**
-     * The ASIO socket to PCIC
-     */
-    boost::asio::ip::tcp::socket sock_;
-
-    /**
-     * The ASIO endpoint to PCIC
-     */
-    boost::asio::ip::tcp::endpoint endpoint_;
-
-    /**
-     * A pointer to the wrapped thread object. This is the thread that
-     * communicates directly with the sensor.
-     */
-    std::unique_ptr<std::thread> thread_;
-
-    /**
-     * Sequential id for callbacks. Used in two-stage mapping:
-     * 1) ticket to callback id
-     * 2) callback id to callback
-     * Using callback ids for cancelling callbacks instead of tickets eliminates
-     * the risk of cancelling a newer callback with the same ticket.
-     */
-    long current_callback_id_;
-
-    /**
-     * Maps PCIC tickets to callback ids. When receiving an incoming message,
-     * the accordant callback id and thus the pending callback can be found via
-     * pending_callbacks_
-     */
-    std::map<int, long> ticket_to_callback_id_;
-
-    /**
-     * Maps callback ids to callbacks. When receiving an incoming message,
-     * the accordant callback can be found (and triggered).
-     */
-    std::map<long, std::function<void(const std::string& content)>>
-      pending_callbacks_;
-
-    /**
-     * Pre-content buffer for incoming messages (<ticket><length>\r\n<ticket>)
-     */
-    std::string in_pre_content_buffer_;
-
-    /**
-     * Content buffer for incoming messages, which is provided
-     * through the callback to the caller
-     */
-    std::string in_content_buffer_;
-
-    /**
-     * Post-content buffer for incoming messages (\r\n)
-     */
-    std::string in_post_content_buffer_;
-
-    /**
-     * Pre-content buffer for outgoing requests (<ticket><length>\r\n<ticket>)
-     */
-    std::string out_pre_content_buffer_;
-
-    /**
-     * Post-content buffer for outgoing messages (\r\n)
-     */
-    std::string out_post_content_buffer_;
-
-    /**
-     * Flag that indicates whether an incoming messages is completely read.
-     */
-    std::atomic_bool out_completed_;
-
-    /**
-     * Ensures synchronized access to ticket and id generation
-     * as well as ticket/id and id/callback maps
-     */
-    std::mutex data_sync_mutex_;
-
-    /**
-     * Ensures single outgoing message
-     */
-    std::mutex out_mutex_;
-
-    /**
-     * Condition variable used to unblock call
-     */
-    std::condition_variable out_cv_;
-
-    /**
-     * Ensures incoming response (used in synchronous Call)
-     */
-    std::mutex in_mutex_;
-
-    /**
-     * Condition variable used to unblock synchronous Call
-     */
-    std::condition_variable in_cv_;
-
+    class Impl;
+    std::unique_ptr<Impl> pImpl;
+   
   }; // end: class PCICClient
 
 } // end: namespace ifm3d
